rgw: Fix Swift API to honor bucket default encryption
Swift uploads bypassed bucket encryption, Swift downloads returned
encrypted garbage for S3-uploaded objects.

- Add get_encrypt_filter/get_decrypt_filter to Swift handlers
- Add decryption support to read_user_manifest_part() for SLO/DLO
- Move get_encryption_defaults() to rgw_crypt.cc (shared by S3+Swift)

Fixes: https://tracker.ceph.com/issues/68525
On-behalf-of: SAP <filipp.akinfiev@clyso.com>
