+++ src/rgw/rgw_rest_swift.cc
@@ -19,6 +19,7 @@
 #include "rgw_formats.h"
 #include "rgw_client_io.h"
 #include "rgw_compression.h"
+#include "rgw_crypt.h"
 
 #include "rgw_auth.h"
 #include "rgw_auth_registry.h"
@@ -1026,6 +1027,13 @@ int RGWPutObj_ObjStore_SWIFT::get_params(optional_yield y)
     chunked_upload = true;
   }
 
+  // Apply bucket-level encryption defaults (same as S3)
+  int ret = get_encryption_defaults(s);
+  if (ret < 0) {
+    ldpp_dout(this, 5) << __func__ << "(): get_encryption_defaults() returned ret=" << ret << dendl;
+    return ret;
+  }
+
   supplied_etag = s->info.env->get("HTTP_ETAG");
   if_match = s->info.env->get("HTTP_IF_MATCH");
   if_nomatch = s->info.env->get("HTTP_IF_NONE_MATCH");
@@ -1153,6 +1161,20 @@ void RGWPutObj_ObjStore_SWIFT::send_response()
   dump_errno(s);
   end_header(s, this);
   rgw_flush_formatter_and_reset(s, s->formatter);
+}
+
+int RGWPutObj_ObjStore_SWIFT::get_encrypt_filter(
+    std::unique_ptr<rgw::sal::DataProcessor> *filter,
+    rgw::sal::DataProcessor *cb)
+{
+  std::unique_ptr<BlockCrypt> block_crypt;
+  int res = rgw_s3_prepare_encrypt(s, s->yield, attrs, &block_crypt,
+                                   crypt_http_responses);
+  if (res == 0 && block_crypt != nullptr) {
+    filter->reset(new RGWPutObj_BlockEncrypt(s, s->cct, cb,
+                                             std::move(block_crypt), s->yield));
+  }
+  return res;
 }
 
 static int get_swift_account_settings(req_state * const s,
@@ -1577,6 +1599,51 @@ int RGWGetObj_ObjStore_SWIFT::get_params(optional_yield y)
   skip_manifest = (mm.compare("get") == 0);
 
   return RGWGetObj_ObjStore::get_params(y);
+}
+
+int RGWGetObj_ObjStore_SWIFT::get_decrypt_filter(
+    std::unique_ptr<RGWGetObj_Filter>* filter,
+    RGWGetObj_Filter* cb,
+    bufferlist* manifest_bl)
+{
+  std::unique_ptr<BlockCrypt> block_crypt;
+  static constexpr bool copy_source = false;
+  int res = rgw_s3_prepare_decrypt(s, s->yield, attrs, &block_crypt,
+                                   &crypt_http_responses, copy_source);
+  if (res < 0) {
+    return res;
+  }
+  if (block_crypt == nullptr) {
+    return 0;
+  }
+
+  // in case of a multipart upload, we need to know the part lengths to
+  // correctly decrypt across part boundaries
+  std::vector<size_t> parts_len;
+
+  // for replicated objects, the original part lengths are preserved in an xattr
+  if (auto i = attrs.find(RGW_ATTR_CRYPT_PARTS); i != attrs.end()) {
+    try {
+      auto p = i->second.cbegin();
+      using ceph::decode;
+      decode(parts_len, p);
+    } catch (const buffer::error&) {
+      ldpp_dout(this, 1) << "failed to decode RGW_ATTR_CRYPT_PARTS" << dendl;
+      return -EIO;
+    }
+  } else if (manifest_bl) {
+    // otherwise, we read the part lengths from the manifest
+    res = RGWGetObj_BlockDecrypt::read_manifest_parts(this, *manifest_bl,
+                                                      parts_len);
+    if (res < 0) {
+      return res;
+    }
+  }
+
+  *filter = std::make_unique<RGWGetObj_BlockDecrypt>(
+      s, s->cct, cb, std::move(block_crypt),
+      std::move(parts_len), s->yield);
+  return 0;
 }
 
 int RGWGetObj_ObjStore_SWIFT::send_response_data_error(optional_yield y)
